Insert table to product.html template page, create deodorant.html
I am uncertain whether or not I should start attempting to code this site
using a framework like React or if I should stick with the jQuery method
of loading in bits of html where necessary. It seems like the learning
curve associated with React is quite steep so it might not be necessary for
just getting something up and running. Since I only have to create 10-20
product pages, I might start by just hardcoding these, at least until I get
a better grasp on what this site will be.